Move HttpSession from impl to core package
Resolves #352
Related to #349

This change will make it easier to re-use this internal Class in the
WebDAV module.
