+++ core/http/src/main/java/org/trellisldp/http/impl/MutatingLdpHandler.java
@@ -58,6 +58,7 @@ import org.trellisldp.api.RuntimeTrellisException;
 import org.trellisldp.api.ServiceBundler;
 import org.trellisldp.api.Session;
 import org.trellisldp.http.core.Digest;
+import org.trellisldp.http.core.HttpSession;
 import org.trellisldp.http.core.TrellisRequest;
 import org.trellisldp.vocabulary.AS;
 import org.trellisldp.vocabulary.LDP;
