+++ core/http/src/test/java/org/trellisldp/http/core/HttpSessionTest.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.trellisldp.http.impl;
+package org.trellisldp.http.core;
 
 import static java.time.Instant.now;
 import static org.junit.jupiter.api.Assertions.*;
